% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataset_info}
\alias{dataset_info}
\title{Dataset directory of the PCAS server}
\format{
A data frame with 56 rows and 8 columns:
\describe{
  \item{Data.type}{Molecular type: Proteome, Phosphoproteome or Transcriptome}
  \item{Dataset}{Study / dataset description}
  \item{Normal}{Number of normal samples (NA = the cohort has no normal samples)}
  \item{Tumor}{Number of tumour samples}
  \item{Abbre}{Dataset abbreviation used in the query functions}
  \item{Study.ID}{Study identifier}
  \item{Disease.Type}{Disease type(s)}
  \item{Primary.Site}{Primary tumour site}
}
}
\source{
PCAS server (\url{https://www.jingege.wang/bioinformatics/PCAS/})
}
\usage{
data("dataset_info")
}
\description{
Table of contents of the CPTAC datasets served by the PCAS API: molecular
type, study, sample counts and the abbreviation (\code{Abbre}) used to query
the data.
}
