% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{drug_info}
\alias{drug_info}
\title{Drug annotation table}
\format{
A data frame with 198 rows and 6 columns:
\describe{
  \item{ID}{Drug id used in the column names of \code{drug_CPTAC}}
  \item{Name}{Drug name}
  \item{Synonyms}{Drug synonyms}
  \item{Targets}{Drug targets}
  \item{Target.pathway}{Signalling pathway of the targets}
  \item{PubCHEM}{PubChem identifier}
}
}
\source{
PCAS server
}
\usage{
data("drug_info")
}
\description{
Drug annotation table
}
