% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ID_list_pro}
\alias{ID_list_pro}
\title{Gene/protein symbols available on the PCAS server}
\format{
A named list.
}
\source{
PCAS server
}
\usage{
data("ID_list_pro")
}
\description{
A named list of identifier sets used to populate gene dropdowns.
}
