% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{idmap_protein}
\alias{idmap_protein}
\title{Protein / phosphosite identifier map}
\format{
A data frame with 172404 rows and 2 columns:
\describe{
  \item{row_names}{Protein accession, or phosphosite accession:site(s)}
  \item{Symbol}{Gene symbol}
}
}
\source{
PCAS server
}
\usage{
data("idmap_protein")
}
\description{
Maps protein ids and phosphorylation-site ids (e.g.
\code{NP_000537.3:s315}, combined ids such as \code{NP_000025.1:s218y223t227})
to gene symbols.
}
