% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TIL_map}
\alias{TIL_map}
\title{Map from infiltration cell types to algorithms}
\format{
A data frame with 137 rows and 2 columns:
\describe{
  \item{cell_type}{Cell-type column name in \code{TIL_CPTAC}}
  \item{algorithm}{Algorithm that produced the score}
}
}
\source{
PCAS server
}
\usage{
data("TIL_map")
}
\description{
Map from infiltration cell types to algorithms
}
