% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uniport_map}
\alias{uniport_map}
\title{Gene symbol to UniProt entry map}
\format{
A data frame with 15162 rows and 4 columns:
\describe{
  \item{From}{UniProt/RefSeq protein accession}
  \item{Entry}{UniProt entry name/accession}
  \item{Reviewed}{"reviewed" when the entry is Swiss-Prot reviewed}
  \item{Symbol}{Gene symbol}
}
}
\source{
UniProt
}
\usage{
data("uniport_map")
}
\description{
Gene symbol to UniProt entry map
}
