% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_cancer_genelist.R
\name{cor_cancer_genelist}
\alias{cor_cancer_genelist}
\title{Perform correlation analysis}
\usage{
cor_cancer_genelist(
  dataset1 = "LUAD_CPTAC_protein",
  id1 = "STAT3",
  dataset2 = "LUAD_CPTAC_mRNA",
  id2 = c("TNS1", "TP53"),
  sample_type = c("Tumor", "Normal"),
  cor_method = "pearson"
)
}
\arguments{
\item{dataset1}{Dataset name. Use 'dataset$Abbre' to get all datasets.}

\item{id1}{Gene symbol, you can input one gene symbols.}

\item{dataset2}{Dataset name. Use 'dataset$Abbre' to get all datasets.}

\item{id2}{Gene symbols, you can input one or multiple symbols.}

\item{sample_type}{Sample type used for correlation analysis, default all types: c("Tumor", "Normal").}

\item{cor_method}{cor_method for correlation analysis, default "pearson".}
}
\description{
Perform correlation analysis of the mRNA/protein expression data in CPTAC database.
}
\examples{
\dontrun{
results <- correlation_analysis(dataset1 = "LUAD_CPTAC_protein",
                                id1 = "STAT3",
                                dataset2 = "LUAD_CPTAC_Phospho",
                                id2 = c("NP_000499.1:s549","NP_000499.1:s551","TP503"),
                                sample_type = c("Tumor","Normal"),
                                cor_method = "pearson")
}
}
