% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_pancancer_genelist.R
\name{cor_pancancer_genelist}
\alias{cor_pancancer_genelist}
\title{Perform correlation analysis}
\usage{
cor_pancancer_genelist(
  df,
  geneset_data,
  sample_type = c("Tumor", "Normal"),
  cor_method = "pearson"
)
}
\arguments{
\item{df}{The expression data of the target gene in multiple datasets, obtained by the get_expr_data() function.}

\item{geneset_data}{The expression data of a genelist in multiple datasets, obtained by the get_expr_data() function.}

\item{sample_type}{Sample type used for correlation analysis, default all types: c("Tumor", "Normal").}

\item{cor_method}{Method for correlation analysis, default "pearson".}
}
\description{
Perform correlation analysis of the mRNA/protein expression data in CPTAC database.
}
\examples{
\dontrun{
genelist <- c("SIRPA","CTLA4","TIGIT","LAG3","VSIR","LILRB2","SIGLEC7","HAVCR2","LILRB4","PDCD1","BTLA")
dataset <- c("CCRCC_CPTAC_protein","GBM_CPTAC_protein","HNSCC_CPTAC_protein","LSCC_CPTAC_protein","LUAD_CPTAC_protein","PDAC_CPTAC_protein","UCEC_CPTAC2_protein","UCEC_CPTAC1_protein")
df <- get_expr_data(genes = "TNS1",datasets = dataset)
geneset_data <- get_expr_data(genes = genelist ,datasets = dataset)
result <- cor_pancancer_genelist(df,geneset_data,sample_type = c("Tumor"))
}
}
