% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_pancancer_TIL.R
\name{cor_pancancer_TIL}
\alias{cor_pancancer_TIL}
\title{Correlation analysis of immune cells infiltration}
\usage{
cor_pancancer_TIL(df, cor_method = "spearman", TIL_type = c("TIMER"))
}
\arguments{
\item{df}{The expression data of the target gene in multiple datasets, obtained by the get_expr_data() function.}

\item{cor_method}{Method for correlation analysis, default "pearson".}

\item{TIL_type}{Algorithm for calculating immune cell infiltration, default "TIMER".}
}
\description{
Calculate the correlation between target gene expression and immune cells infiltration in multiple datasets.
}
\examples{
\dontrun{
dataset <- c("CCRCC_CPTAC_protein","GBM_CPTAC_protein","HNSCC_CPTAC_protein","LSCC_CPTAC_protein","LUAD_CPTAC_protein","PDAC_CPTAC_protein","UCEC_CPTAC2_protein","UCEC_CPTAC1_protein")
df <- get_expr_data(genes = "TNS1",datasets = dataset)
result <- cor_pancancer_TIL(df,Target.pathway = c("Cell cycle"))
}
}
