% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_expr_data.R
\name{get_expr_data}
\alias{get_expr_data}
\title{Get CPTAC expression data}
\usage{
get_expr_data(
  datasets = c("LUAD_CPTAC_protein", "LSCC_CPTAC_protein"),
  genes = c("TP53", "TNS1")
)
}
\arguments{
\item{datasets}{Dataset names, you can input one or multiple datasets. Use 'dataset$Abbre' to get all datasets.}

\item{genes}{Gene symbols, you can input one or multiple symbols.}
}
\description{
Get the mRNA/protein expression data in CPTAC database.
}
\examples{
\dontrun{
results <- get_expr_data(datasets = "LUAD_CPTAC_mRNA", genes = c("GAPDH","TNS1"))
results <- get_expr_data(datasets = c("LUAD_CPTAC_protein","LSCC_CPTAC_protein"), genes = "GAPDH")
results <- get_expr_data(datasets = c("CCRCC_CPTAC_mRNA","GBM_CPTAC_mRNA","HNSCC_CPTAC_mRNA","LSCC_CPTAC_mRNA","LUAD_CPTAC_mRNA","PDAC_CPTAC_mRNA","UCEC_CPTAC2_mRNA","UCEC_CPTAC1_mRNA"), genes = c("SIRPA","CTLA4","TIGIT","LAG3","VSIR","LILRB2","SIGLEC7","HAVCR2","LILRB4","PDCD1","BTLA"))
}
}
