% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_cor_heatmap.R
\name{viz_cor_heatmap}
\alias{viz_cor_heatmap}
\title{Visualization of correlation results}
\usage{
viz_cor_heatmap(r, p)
}
\arguments{
\item{r}{The correlation coefficient matrix r of the correlation analysis results obtained from the functions cor_pancancer_genelist(), cor_pancancer_TIL(), and cor_pancancer_drug().}

\item{p}{The P-value matrix p of the correlation analysis results obtained from the functions cor_pancancer_genelist(), cor_pancancer_TIL(), and cor_pancancer_drug().}
}
\description{
Presenting correlation analysis results using heat maps based on ggplot2.
}
\examples{
\dontrun{
genelist <- c("SIRPA","CTLA4","TIGIT","LAG3","VSIR","LILRB2","SIGLEC7","HAVCR2","LILRB4","PDCD1","BTLA")
dataset <- c("CCRCC_CPTAC_mRNA","GBM_CPTAC_mRNA","HNSCC_CPTAC_mRNA","LSCC_CPTAC_mRNA","LUAD_CPTAC_mRNA","PDAC_CPTAC_mRNA","UCEC_CPTAC2_mRNA","UCEC_CPTAC1_mRNA")
df <- get_expr_data(genes = "TNS1",datasets = dataset)
geneset_data <- get_expr_data(genes = genelist ,datasets = dataset)
result <- cor_pancancer_genelist(df,geneset_data,sample_type = c("Tumor"))
viz_cor_heatmap(result$r,result$p)
}
}
