% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_corplot.R
\name{viz_corplot}
\alias{viz_corplot}
\title{Plotting scatter plot}
\usage{
viz_corplot(
  data,
  a,
  b,
  method = "pearson",
  x_lab = " relative expression",
  y_lab = " relative expression"
)
}
\arguments{
\item{data}{A gene expression dataset with at least two genes included, rows represent samples, and columns represent gene expression in the matrix.}

\item{a}{Gene A}

\item{b}{Gene B}

\item{method}{Method for correlation analysis, "pearson" or "spearman".}

\item{x_lab}{X-axis label.}

\item{y_lab}{Y-axis label.}
}
\description{
Scatter plot with sample size (n), correlation coefficient (r) and p value (p.value).
}
