% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_DEGs_volcano.R
\name{viz_DEGs_volcano}
\alias{viz_DEGs_volcano}
\title{Volcano plot for DEGs}
\usage{
viz_DEGs_volcano(
  df,
  p.cut = 0.05,
  logFC.cut = 1,
  show.top = FALSE,
  show.labels = NULL
)
}
\arguments{
\item{cohort}{Data cohort, for example, "LUAD_APOLLO", "LUAD_CPTAC".}

\item{data_input}{Expression data obtained from get_expr_data() function.}
}
\description{
Plotting volcano plot for DEGs between tumor and normal samples in CPTAC datasets.
}
\examples{
\dontrun{
results <- get_DEGs_result(dataset = "LUAD_CPTAC_protein", method = "limma")
results <- get_DEGs_result(dataset = "LUAD_CPTAC_mRNA", method = "t.test")
viz_DEGs_volcano(results)
}
}
