% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_phoso_sites.R
\name{viz_phoso_sites}
\alias{viz_phoso_sites}
\title{Phosphorylation site information}
\usage{
viz_phoso_sites(gene = "YTHDC2", phoso_infoDB = "CPTAC")
}
\arguments{
\item{gene}{Gene/protein symbol.}

\item{phoso_infoDB}{Database for extracting phosphorylation site information. Default "CPTAC".}
}
\description{
Query phosphorylation site information of target proteins based on CPTAC database phosphorylation proteomics data or UniProt database.
}
\examples{
\dontrun{
viz_phoso_sites("TNS1")
viz_phoso_sites("YTHDC2",phoso_infoDB= "UniProt")
}
}
