% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_TvsN.R
\name{viz_TvsN}
\alias{viz_TvsN}
\title{Visualizing genes expression}
\usage{
viz_TvsN(
  df,
  df_type = c("single", "multi_gene", "multi_set"),
  Show.P.value = TRUE,
  Show.P.label = TRUE,
  Method = "t.test",
  values = c("#00AFBB", "#FC4E07"),
  Show.n = TRUE,
  Show.n.location = "default"
)
}
\arguments{
\item{df}{Gene expression data obtained from get_expr_data().}

\item{df_type}{The type of gene expression data, one value of "single","multi_gene", and "multi_set".}

\item{Show.P.value}{Whether to display the results of differential analysis, default TRUE.}

\item{Show.P.label}{Whether to display significance markers for differential analysis, default TRUE.}

\item{Method}{Methods of differential analysis, "t.test" or "limma", default "t.test".}

\item{values}{Color palette for normal and tumor groups. Default c("#00AFBB", "#FC4E07").}

\item{Show.n}{Display sample size.}

\item{Show.n.location}{Y-axis position displayed for sample size.}
}
\description{
Visualizing the different expression of mRNA/protein expression data between Tumor and Normal tissues in CPTAC database.
}
\examples{
\dontrun{
df_single <- get_expr_data(datasets = "LUAD_CPTAC_mRNA",genes = c("TP53"))
df_multi_gene <- get_expr_data(datasets = "LUAD_CPTAC_protein",genes = c("TP53","TNS1"))
df_multi_set <- get_expr_data(datasets=c("CCRCC_CPTAC_protein","GBM_CPTAC_protein","HNSCC_CPTAC_protein","LSCC_CPTAC_protein","LUAD_CPTAC_protein" ,
                                         "PDAC_CPTAC_protein" , "UCEC_CPTAC2_protein", "UCEC_CPTAC1_protein"),genes = c("TNS1"))
viz_TvsN(df_single,df_type = "single")
viz_TvsN(df_multi_gene,df_type = "multi_gene")
viz_TvsN(df_multi_set,df_type = "multi_set")
}
}
